+++ src/App.js
@@ -1,4 +1,4 @@
-import Home from './pages/Home';
+import Home from './pages/Home/Home';
 import { createGlobalStyle } from 'styled-components'
